SD---Taxis/requirements.txt
colorama
kafka-python
numpy
requests
pyjwt
mysql-connector-python
